requirements.txt
fastapi[all]
mysql-connector-python
pymysql
sqlalchemy

